This functions as far as provisioning new hosts, re-provisioning any that have old versions of the config or JAR and determining which hosts are ready for work.
The PrimaryDaemon does not loop yet, the WorkerDaemon does and continues running on a node until it determines that it is stale.

List of worker hosts can be changed during run of Primary without need to restart.

TODO
Primary Daemon needs to know what to do if a worker reports in when not in the hosts file.
Worker needs actual docker monitoring thread